requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
sqlalchemy==2.0.36
aiomysql==0.2.0
mysql-connector-python==9.1.0
google-api-python-client==2.154.0
google-auth==2.37.0
python-dotenv==1.0.1
pydantic==2.10.3